src/main/java/model/Department.java
package model;

public class Department {
    private int id;
    private String deptName;
    private String deptDescription;
    private int employeeCount;

    public Department(String deptName, String deptDescription, int employeeCount) {
        this.deptName = deptName;
        this.deptDescription = deptDescription;
        this.employeeCount = employeeCount;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getDeptName() {
        return deptName;
    }

    public void setDeptName(String deptName) {
        this.deptName = deptName;
    }

    public String getDeptDescription() {
        return deptDescription;
    }

    public void setDeptDescription(String deptDescription) {
        this.deptDescription = deptDescription;
    }

    public int getEmployeeCount() {
        return employeeCount;
    }

    public void setEmployeeCount(int employeeCount) {
        this.employeeCount = employeeCount;
    }
}
